requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
edge-tts==7.2.3
fastapi==0.104.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
loguru==0.7.2
multidict==6.6.4
propcache==0.4.0
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
python-dotenv==1.0.0
python-telegram-bot==20.7
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.21.0
chromadb>=1.1.1
sentence-transformers>=5.1.0
